requirements.txt
pandas==2.2.2
matplotlib==3.9.2
jinja2==3.1.4
python-dateutil==2.9.0.post0